# docs/bravo/openapi.yaml
openapi: 3.0.3
info:
  title: SilentApply API
  version: 1.0.0
  description: |
    Thin, explicit, correlation-aware API for SilentApply.

    All endpoints support correlation tracking via the X-Correlation-ID header.
    Format: t:<tenant>|c:<uuidv7>

    Canon-compliant: calm, bounded, candidate-controlled.
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: https://api.silentapply.com
    description: Production
  - url: http://localhost:3000
    description: Local development
security:
  - sessionAuth: []
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: NextAuth session cookie (required for authenticated endpoints)
  parameters:
    CorrelationHeader:
      name: X-Correlation-ID
      in: header
      required: false
      schema:
        type: string
        pattern: '^t:[a-zA-Z0-9_-]+\|c:[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-7[0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}$'
      description: |
        Correlation ID format: t:<tenant>|c:<uuidv7>
        Generated server-side if not provided.
        Propagated to OMEGA tasks and all logs.
  schemas:
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          example: BAD_REQUEST
        message:
          type: string
          example: Profile handle already exists
        correlationId:
          type: string
          example: t:default|c:01932a3c-4b5e-7890-abcd-ef1234567890
...
    ResumeUploadResponse:
      type: object
      required:
        - resumeId
        - status
      properties:
        resumeId:
          type: string
          format: uuid
        status:
          type: string
          enum: [uploaded, processing]
        message:
          type: string
          example: Resume uploaded. Ingestion queued.
...
paths:
...
  /api/resume:
    post:
      summary: Upload resume
      description: |
        Uploads a resume file. Triggers OMEGA task: resume.ingest
        Accepts PDF, DOCX (max 5MB).
      operationId: uploadResume
      security:
        - sessionAuth: []
      parameters:
        - $ref: '#/components/parameters/CorrelationHeader'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '201':
          description: Resume uploaded, ingestion queued
          headers:
            X-Correlation-ID:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResumeUploadResponse'
        '400':
          description: Invalid file format or size
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate limit exceeded (quiet)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      tags:
        - Resume
...
  /api/qa:
    post:
      summary: Submit a question (public, no auth)
      description: |
        Submits a recruiter question. Triggers OMEGA task: qa.answer
        Rate limited quietly.
      operationId: submitQuestion
      security: []
      parameters:
        - $ref: '#/components/parameters/CorrelationHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - profileId
                - question
              properties:
                profileId:
                  type: string
                  format: uuid
                question:
                  type: string
                  maxLength: 1000
                email:
                  type: string
                  format: email
                  nullable: true
      responses:
        '202':
          description: Question received, answer queued
          headers:
            X-Correlation-ID:
              schema:
                type: string
          content:
            application/json:
              schema:
                type: object
                properties:
                  threadId:
                    type: string
                    format: uuid
                  status:
                    type: string
                    example: processing
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Profile not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate limit exceeded (quiet)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      tags:
        - Public
...